rag/requirements.txt
openai
numpy
python-dotenv
black
wikipedia-api
pinecone
requests